+++ DungeonDrive/DungeonDrive/WeaponStats.cs
@@ -11,7 +11,7 @@ namespace DungeonDrive
         public static int[] atk_damage = { 1, 3 };
         public static double[] atk_speed = { 0.1, 0.2 };
         public static double[] proj_speed = { 0.1, 0.2 };
-        public static double[] range = { 5, 12 };
+        public static int[] range = { 5, 12 };
         public static double[] slowSec = { 0.5, 2 };
         public static double[] slowFac = { 0.3, 0.5 };
     }
